Re-arrange segments migrations
The 1st migration won't apply because it wants to make `segmentId` part
of primary key. But it's not possible until we fill the field with data.
We only fill it with data in the 2nd migration.

Since we remove the field in the 3rd migration, lets just not creat it
in the first place