Fix my-project-groups query
The `foundation` sub-query returns only project-groups, but admin
segments only contain sub-projects. So it doesn't make much sense to
filter the results of one with the other ones.

Instead we use admin segments to filter inside the sub-query, where we
we still have access to sub-projects.

This way we still in the end return project-groups, but only those which
include sub-projects, which the user can manage as a project-admin.