Add a middleware to backed which stores user-data in the request
It would be convenient to just access these `ip` and `headers`
propertiers further down the road in audit-logs, but because we copy
fields of the original `req` object into `IServiceOptions` and
`IRepoOptions` like so:
   {
      ...req,
      transaction: ...
   }
the `req` object loses its prototype together with accessor methods the
object had initially. This means the raw data is there, but convenient
data accessors are lost.

So instead of reimplementing those accessors, we just pre-save necessary
information in fields, so they're available even when the prototype is
lost.