Reduce number of partitions per topic from 10 to 2
Right now we have a total limit of 15 partitions for the whole tenant in
Oracle Cloud. I have a feeling, that going with 10 partitions for each
queue is a bit of an overkill for us, so we bring it down to fit in the
limit, and also have some room for new potential queues.