fix: don't initialise QuestDB automatically (#3237)
Due to a top-level variable being initialised in the data access layer, it was being executed immediately when the code was imported, which was causing issues when QuestDB was not in use.

This moves the variable initialistion to inside the function that uses it, so we don't have the above problem.

Signed-off-by: Raúl Santos <4837+borfast@users.noreply.github.com>